feat: Refactor flyout menu and move delete credential to teleport
closes #1280

closes #1184
